Move host types into runtime and remove deprecated host package
