+++ packages/client/tests/smoke.spec.ts
@@ -1,6 +1,37 @@
 import { test, expect } from "@playwright/test";
-import { generateEmbedURL } from "@runno/host";
 import type { TerminalElement } from "@runno/runtime";
+import { encode } from "url-safe-base64";
+
+function generateEmbedURL(
+  code: string,
+  runtime: string,
+  options?: {
+    showControls?: boolean; // Default: true
+    showEditor?: boolean; // Default: true
+    autorun?: boolean; // Default: false
+    baseUrl?: string; // Default: "https://runno.run/"
+  }
+): URL {
+  const showEditor =
+    options?.showEditor === undefined ? true : options?.showEditor;
+  const showControls =
+    options?.showControls === undefined ? true : options?.showControls;
+  const autorun = options?.autorun || false;
+  const baseUrl = options?.baseUrl || "https://runno.run/";
+  const url = new URL(baseUrl);
+  if (showEditor) {
+    url.searchParams.append("editor", "1");
+  }
+  if (autorun) {
+    url.searchParams.append("autorun", "1");
+  }
+  if (!showControls) {
+    url.searchParams.append("controls", "0");
+  }
+  url.searchParams.append("runtime", runtime);
+  url.searchParams.append("code", encode(btoa(code)));
+  return url;
+}
 
 const baseURL = "http://localhost:5678/";
