--- packages/host/.gitignore
@@ -1,3 +1,0 @@
-*.log
-node_modules
-dist
