--- packages/host/package.json
@@ -1,42 +1,0 @@
-{
-  "name": "@runno/host",
-  "version": "0.5.4",
-  "author": "Ben Taylor <runno@taybenlor.com>",
-  "description": "DEPRECATED iframe host for @runno",
-  "website": "https://runno.dev",
-  "private": true,
-  "repository": {
-    "type": "git",
-    "url": "git://github.com/taybenlor/runno.git"
-  },
-  "license": "MIT",
-  "publishConfig": {
-    "access": "public"
-  },
-  "main": "dist/index.js",
-  "module": "dist/index.esm.js",
-  "types": "dist/index.d.ts",
-  "files": [
-    "dist",
-    "src"
-  ],
-  "engines": {
-    "node": ">=12"
-  },
-  "scripts": {
-    "build:esm": "esbuild src/index.ts --bundle --format=esm --outfile=dist/index.esm.js",
-    "build:cjs": "esbuild src/index.ts --bundle --format=cjs --outfile=dist/index.js",
-    "build:types": "tsc",
-    "build": "npm run build:esm && npm run build:cjs && npm run build:types"
-  },
-  "devDependencies": {
-    "@types/node": "^16.11.4",
-    "@types/url-safe-base64": "^1.1.0",
-    "esbuild": "^0.14.48",
-    "post-me": "^0.4.5",
-    "tslib": "^2.2.0",
-    "typescript": "^4.4.4",
-    "url-safe-base64": "^1.1.1"
-  },
-  "gitHead": "812d21ec7c0c5d57c0274caaf7ff21c79ca79c1f"
-}
