+++ packages/runtime/lib/types.ts
@@ -1,3 +1,5 @@
+import type { WASIFS } from "@runno/wasi";
+
 export type Runtime =
   | "python"
   | "quickjs"
@@ -39,48 +41,6 @@ export type TerminatedResult = {
 };
 
 export type RunResult = CompleteResult | CrashResult | TerminatedResult;
-
-export type WASIPath = string;
-
-export type WASIFS = {
-  [path: WASIPath]: WASIFile;
-};
-
-export type BinaryWASIFS = {
-  [path: WASIPath]: WASIBinaryFile;
-};
-
-export type WASITimestamps = {
-  access: Date;
-  modification: Date;
-  change: Date;
-};
-
-export type WASIFile = {
-  path: WASIPath; // TODO: This duplication is annoying, lets remove it
-  timestamps: WASITimestamps;
-} & (
-  | {
-      mode: "string";
-      content: string;
-    }
-  | {
-      mode: "binary";
-      content: Uint8Array;
-    }
-);
-
-export type WASIBinaryFile = {
-  path: WASIPath;
-  timestamps: WASITimestamps;
-  mode: "binary";
-  content: Uint8Array;
-};
-
-export type WASIExecutionResult = {
-  exitCode: number;
-  fs: WASIFS;
-};
 
 export type RuntimeMethods = {
   showControls: () => void;
